rc file format change (please read below)
First: if you have an old rc file it will still work.  In fact
internally the new rc file data gets converted to the old one.  So you
do not have to do anything normally.

This changes the rc file in the following way, taking mirroring as an
example.  Simple enough for most purposes but you really should read the
documentation...

old:
    INPUT                       =>
        [
            'Mirroring::input',
        ],

    PRE_GIT                     =>
        [
            'Mirroring::pre_git',
        ],

    POST_GIT                    =>
        [
            'Mirroring::post_git',
        ],

new:
    ENABLE  =>
        [
        # COMMANDS
            ... several commands ...
        # Triggers
            ... several triggers ...
            'Mirroring',
        ],